+++ Tests/PropTypes/IsRequired.js
@@ -1,28 +1,28 @@
-import nodeProto from './../../Dist/NodeProto.js';
+var nodeProto = require('./../../Src/NodeProto.js');
 
 const propType = nodeProto.propTypes.isRequired;
 
-describe('NodeProto: propTypes.isRequired', function () {
-    it('should be defined', function () {
+describe('NodeProto: propTypes.isRequired', () => {
+    it('should be defined', () => {
         expect(propType).toBeDefined();
     });
 
-    it('should return an object containing a result and value when called.', function () {
+    it('should return an object containing a result and value when called.', () => {
         expect(propType()).toEqual({
-            result: 0,
+            result: false,
             value: undefined
         });
     });
 
-    it('should return a positive result if an argument was passed.', function () {
+    it('should return a positive result if an argument was passed.', () => {
         expect(propType('Prop').result).toBeTruthy();
     });
 
-    it('should return a positive result even if the argument passed argument is "false".', function () {
+    it('should return a positive result even if the argument passed argument is "false".', () => {
         expect(propType(false).result).toBeTruthy();
     });
 
-    it('should return the passed argument as the returning value', function () {
+    it('should return the passed argument as the returning value.', () => {
         expect(propType('Prop').value).toBe('Prop');
     });
 });